meal-grocery-planner-crewai/requirements.txt
# Meal & Grocery Planner with CrewAI
# Python dependencies

# Core AI frameworks
crewai==0.141.0
langchain==0.3.20
langchain-community==0.3.19
langchain-openai==0.3.25

# Tools and utilities
crewai-tools==0.51.1
duckduckgo-search==7.5.2

# Data validation
pydantic>=2.0.0

# LLM providers
openai>=1.0.0
ibm-watson>=5.0.0

# Web search
serper-python>=0.0.5

# Utilities
python-dotenv>=1.0.0
ipython>=8.0.0